requirements.txt
flask>=3.0
gunicorn>=22.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv>=1.0
playwright>=1.40
